Fix ruby build failure on the latest archlinux:base-devel image (#1)
In summary, we first switch to my fork of ` in jeffreytse/jekyll-deploy-action` and fix the compiling error of `ruby<3.2` on arch linux; then we upgrade `github-pages` gem and resolve the `untaint` issue that occurs for `ruby>=3.2`.

---

* Try upgrading to ruby 3.4.3

Because ruby 3.1.3 build failed on the latest archlinux:base-devel docker image, which is used in jeffreytse/jekyll-deploy-action@v0.6.0 github action.

* Try ruby 3.3.0

* Try ruby 3.2.0

* Try my fork of jekyll-deploy-action

* Try upgrading `github-pages` gem

Because we encountered the jekyll bug https://github.com/jekyll/jekyll/issues/9233. Hopefully by upgrading the `github-pages` gem we are able to resolve it. Otherwise, we will need to rollback ruby version to 3.1.3.